genesis_postit/genesis_api/serializers.py
from rest_framework import serializers
from . import models

class AlbumReviewSerializer(serializers.ModelSerializer):
    user = serializers.ReadOnlyField(source="user.username")
    user_id = serializers.ReadOnlyField(source="user.id")