requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
Django==4.1.3
djangorestframework==3.14.0
entrypoints==0.4
executing==1.0.0
greenlet==1.1.3.post0
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pywin32==304
pyzmq==23.2.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
sqlparse==0.4.3
stack-data==0.5.0
tornado==6.2
traitlets==5.3.0
tzdata==2022.6
wcwidth==0.2.5
